requirements.txt
# Telegram Bot
python-telegram-bot>=21.5

# MongoDB
motor>=3.6.0

# Authentication
bcrypt>=4.2.1
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Web Framework  
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Utilities
pydantic
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.20
aiofiles>=24.1.0
httpx>=0.28.0

# QR Code Generation
qrcode>=7.4.0
Pillow>=11.0.0

langchain_mistralai
python-telegram-bot[webhooks]